Initialize git repo with misc files.
The assembler is currently able to do an inefficient match that
resembles instruction description.  Furthermore, there is a scetch of
opcode encoding based on intel's opcode discription format.

More work needs to be done in order to specify the layout of the
project, short term goals, and long term goals.